+++ services/frontend/src/api/index.ts
@@ -1,31 +1,32 @@
 import { mockEventsApi } from './mock-services/events'
 import { createEventsApi } from './services/events'
-import { mockMediaApi } from './mock-services/media'
-import { createMediaApi } from './services/media'
+import { mediaApi } from './services/media'
 import { mockFeedApi } from './mock-services/feed'
 import { createFeedApi } from './services/feed'
 import { mockInteractionsApi } from './mock-services/interactions'
-import { createInteractionsApi } from './services/interactions'
+// import { createInteractionsApi } from './services/interactions'
 import { mockUsersApi } from './mock-services/users'
-import { createUsersApi } from './services/users'
+// import { createUsersApi } from './services/users'
 import {
   createEventsClient,
-  createFeedClient,
-  createInteractionsClient,
-  createMediaClient,
-  createUsersClient,
+  // createFeedClient,
+  // createInteractionsClient,
+  // createUsersClient,
 } from './client'
 
-const useMockApi = import.meta.env.VITE_USE_MOCK_API
-  ? import.meta.env.VITE_USE_MOCK_API === 'true'
-  : import.meta.env.DEV
+const useRealApi: boolean = import.meta.env.VITE_USE_MOCK_API === 'false'
+console.log('Using real API:', useRealApi)
 
 export const api = {
-  events: useMockApi ? mockEventsApi : createEventsApi(createEventsClient()),
-  media: useMockApi ? mockMediaApi : createMediaApi(createMediaClient()),
-  feed: useMockApi ? mockFeedApi : createFeedApi(createFeedClient()),
-  interactions: useMockApi
-    ? mockInteractionsApi
-    : createInteractionsApi(createInteractionsClient()),
-  users: useMockApi ? mockUsersApi : createUsersApi(createUsersClient()),
+  events: useRealApi ? createEventsApi(createEventsClient()) : mockEventsApi,
+  media: mediaApi,
+  feed: useRealApi ? createFeedApi(createEventsClient()) : mockFeedApi,
+  interactions: mockInteractionsApi,
+  users: mockUsersApi,
+  // media: useMockApi ? mockMediaApi : createMediaApi(createMediaClient()),
+  // feed: useMockApi ? mockFeedApi : createFeedApi(createFeedClient()),
+  // interactions: useMockApi
+  //   ? mockInteractionsApi
+  //   : createInteractionsApi(createInteractionsClient()),
+  // users: useMockApi ? mockUsersApi : createUsersApi(createUsersClient()),
 }