+++ services/frontend/src/api/interfaces/events.ts
@@ -1,5 +1,6 @@
-import type { Tag, Event, EventID, EventData, EventStatus } from '../types/events'
+import type { Tag, Event, EventID, PartialEventData, EventStatus } from '../types/events'
 import type { UserID } from '../types/users'
+import type { PaginatedRequest, PaginatedResponse } from './commons'
 
 export interface TagCategory {
   category: string
@@ -8,23 +9,35 @@ export interface TagCategory {
 
 export type GetTagResponse = TagCategory[]
 
-export interface GetEventByIdResponse {
-  event: Event
-}
+export type GetEventByIdResponse = Event
 
 export interface PublishEventResponse {
-  eventId: EventID
+  id_event: EventID
 }
 
 export interface EventsDataResponse {
   events: Event[]
 }
 
+export interface EventPaginatedResponse {
+  events: Event[]
+  limit: number
+  offset: number
+  hasMore: boolean
+}
+
 export interface EventAPI {
   getTags(): Promise<GetTagResponse>
-  getEventById(id: EventID): Promise<GetEventByIdResponse>
-  getEventsByIds(ids: EventID[]): Promise<EventsDataResponse>
-  publishEvent(eventData: EventData): Promise<PublishEventResponse>
-  searchByName(query: string): Promise<EventsDataResponse>
-  getEventsByUserIdAndStatus(userId: UserID, status: EventStatus): Promise<EventsDataResponse>
+  getEventById(id_event: EventID): Promise<GetEventByIdResponse>
+  getEventsByIds(id_events: EventID[]): Promise<EventsDataResponse>
+  publishEvent(eventData: PartialEventData): Promise<PublishEventResponse>
+  updateEventData(id_event: EventID, eventData: PartialEventData): Promise<void>
+  updateEventPoster(id_event: EventID, poster: File): Promise<void>
+  deleteEvent(id_event: EventID): Promise<void>
+  searchByName(title: string, pagination?: PaginatedRequest): Promise<PaginatedResponse<Event>>
+  getEventsByUserIdAndStatus(
+    id_organization: UserID,
+    status: EventStatus,
+    pagination?: PaginatedRequest
+  ): Promise<PaginatedResponse<Event>>
 }