+++ services/frontend/src/api/services/events.ts
@@ -4,29 +4,80 @@ import type {
   GetEventByIdResponse,
   PublishEventResponse,
   EventsDataResponse,
+  EventPaginatedResponse,
 } from '../interfaces/events'
 import type { GetTagResponse } from '../interfaces/events'
-import type { EventID, EventData } from '../types/events'
-import { buildQueryParams } from '../utils'
+import type { EventID, PartialEventData, Event } from '../types/events'
+import { buildQueryParams, evaluatePagination } from '../utils'
+import type { PaginatedRequest, PaginatedResponse } from '../interfaces/commons'
 
 export const createEventsApi = (eventsClient: ApiClient): EventAPI => ({
   async getTags(): Promise<GetTagResponse> {
     return eventsClient.get<GetTagResponse>('/tags')
   },
-  async getEventById(id: EventID): Promise<GetEventByIdResponse> {
-    return eventsClient.get<GetEventByIdResponse>(`/events/${id}`)
+  async getEventById(id_event: EventID): Promise<GetEventByIdResponse> {
+    return eventsClient.get<GetEventByIdResponse>(`/${id_event}`)
   },
-  async getEventsByIds(ids: EventID[]): Promise<EventsDataResponse> {
-    const eventsResponses = await Promise.all(ids.map((id) => this.getEventById(id)))
-    return { events: eventsResponses.map((r) => r.event) }
+  async getEventsByIds(id_events: EventID[]): Promise<EventsDataResponse> {
+    const eventsResponses = await Promise.all(
+      id_events.map((id_event) => this.getEventById(id_event))
+    )
+    return { events: eventsResponses }
   },
-  async publishEvent(eventData: EventData): Promise<PublishEventResponse> {
-    return eventsClient.post<PublishEventResponse>('/events', eventData)
+  async publishEvent(eventData: PartialEventData): Promise<PublishEventResponse> {
+    const { poster, date, ...rest } = eventData
+    const formData = new FormData()
+    if (poster) {
+      formData.append('poster', poster)
+    }
+    const backendEventData = {
+      ...rest,
+      date: date?.toISOString().replace(/\.\d{3}Z$/, ''),
+    }
+    console.log('publishEvent backendEventData', backendEventData)
+    formData.append('event', JSON.stringify(backendEventData))
+    return eventsClient.post<{ id_event: string }>('/', formData)
   },
-  async searchByName(query: string): Promise<EventsDataResponse> {
-    return eventsClient.get<EventsDataResponse>(`/events${buildQueryParams({ query })}`)
+  async updateEventData(id_event: EventID, eventData: PartialEventData): Promise<void> {
+    const { poster, date, ...rest } = eventData
+    const backendEventData = {
+      ...rest,
+      date: date?.toISOString().replace(/\.\d{3}Z$/, ''),
+    }
+    console.log('updateEventData backendEventData', backendEventData)
+    await eventsClient.put(`/${id_event}`, backendEventData)
   },
-  async getEventsByUserIdAndStatus(userId: string, status: string): Promise<EventsDataResponse> {
-    return eventsClient.get<EventsDataResponse>(`/events${buildQueryParams({ userId, status })}`)
+  async updateEventPoster(id_event: EventID, poster: File): Promise<void> {
+    const formData = new FormData()
+    formData.append('poster', poster)
+    await eventsClient.post(`/${id_event}/poster`, formData)
+  },
+  async deleteEvent(id_event: EventID): Promise<void> {
+    await eventsClient.delete(`/${id_event}`)
+  },
+  async searchByName(
+    id_organization: string,
+    pagination?: PaginatedRequest
+  ): Promise<PaginatedResponse<Event>> {
+    const { events: items, ...rest } = await eventsClient.get<EventPaginatedResponse>(
+      `/search${buildQueryParams({ id_organization, ...evaluatePagination(pagination) })}`
+    )
+    return {
+      items,
+      ...rest,
+    }
+  },
+  async getEventsByUserIdAndStatus(
+    id_organization: string,
+    status: string,
+    pagination?: PaginatedRequest
+  ): Promise<PaginatedResponse<Event>> {
+    const { events: items, ...rest } = await eventsClient.get<EventPaginatedResponse>(
+      `/search${buildQueryParams({ id_organization, status, ...evaluatePagination(pagination) })}`
+    )
+    return {
+      items,
+      ...rest,
+    }
   },
 })